flask_ui/requirements.txt
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap4==4.0.2
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
